Last commit info on Info level, full details on debug
