+++ livingatlas/pipelines/src/main/java/au/org/ala/pipelines/beam/ALAUUIDMintingPipeline.java
@@ -2,6 +2,7 @@ package au.org.ala.pipelines.beam;
 
 import au.org.ala.kvs.ALAPipelinesConfig;
 import au.org.ala.kvs.ALAPipelinesConfigFactory;
+import au.org.ala.pipelines.util.VersionInfo;
 import au.org.ala.kvs.cache.ALAAttributionKVStoreFactory;
 import au.org.ala.kvs.client.ALACollectoryMetadata;
 import au.org.ala.pipelines.common.ALARecordTypes;
@@ -73,6 +74,7 @@ public class ALAUUIDMintingPipeline {
   public static final String UNIQUE_COMPOSITE_KEY_JOIN_CHAR = "|";
 
   public static void main(String[] args) throws Exception {
+    VersionInfo.print();
     String[] combinedArgs = new CombinedYamlConfiguration(args).toArgs("general", "uuid");
     UUIDPipelineOptions options =
         PipelinesOptionsFactory.create(UUIDPipelineOptions.class, combinedArgs);
